move LabelSelectorAsSelector out of mesh pkg, pilot-agent needs to be light but imports this package
